extract effect style work
